+++ go/core/pkg/sandboxbackend/openshell/openshell.go
@@ -4,9 +4,6 @@
 // Use Dial to obtain OpenShellClients (shared connection for openshell.v1.OpenShell
 // and openshell.inference.v1.Inference).
 //
-// • NewOpenShellBackend — generic AgentHarness resources with spec.backend=openshell:
-// user image/env mapping per translate.go; spec.modelConfigRef is ignored for pre-create and bootstrap.
-//
 // • NewOpenClawBackend — pin the sandbox image to NemoclawSandboxBaseImage, translateModelConfig
 // when modelConfigRef is set, run OpenClaw bootstrap after Ready. The same instance
 // is registered for spec.backend=openclaw and nemoclaw (see app wiring).
@@ -29,19 +26,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-// PreCreateSandboxFunc is an optional hook: runs after GetSandbox returns NotFound and before
-// CreateSandbox on the OpenShell-backed control plane. Each AgentHarness backend passes nil or a
-// backend-specific function (e.g. OpenClaw uses translateModelConfig to apply spec.modelConfigRef; a
-// future Hermes backend would pass its own implementation). A separate interface would be a single-method
-// duplicate of this func type, so the hook stays a function.
-type PreCreateSandboxFunc func(ctx context.Context, ah *v1alpha2.AgentHarness, kube client.Client, oc *OpenShellClients) error
-
 type agentHarnessOpenShellBackend struct {
 	*AgentHarnessOpenShellClient
-	kubeClient       client.Client
-	backendName      v1alpha2.AgentHarnessBackendType
-	buildCreate      func(*v1alpha2.AgentHarness) (*openshellv1.CreateSandboxRequest, []string)
-	preCreateSandbox PreCreateSandboxFunc
+	kubeClient  client.Client
+	backendName v1alpha2.AgentHarnessBackendType
+	buildCreate func(*v1alpha2.AgentHarness) (*openshellv1.CreateSandboxRequest, []string)
 }
 
 func newAgentHarnessOpenShellBackend(
@@ -51,14 +40,12 @@ func newAgentHarnessOpenShellBackend(
 	recorder record.EventRecorder,
 	name v1alpha2.AgentHarnessBackendType,
 	build func(*v1alpha2.AgentHarness) (*openshellv1.CreateSandboxRequest, []string),
-	preCreateSandbox PreCreateSandboxFunc,
 ) *agentHarnessOpenShellBackend {
 	return &agentHarnessOpenShellBackend{
 		AgentHarnessOpenShellClient: newAgentHarnessOpenShellClient(clients, cfg, recorder),
 		kubeClient:                  kubeClient,
 		backendName:                 name,
 		buildCreate:                 build,
-		preCreateSandbox:            preCreateSandbox,
 	}
 }
 
@@ -67,41 +54,35 @@ func (b *agentHarnessOpenShellBackend) Name() v1alpha2.AgentHarnessBackendType {
 	return b.backendName
 }
 
-// EnsureAgentHarness implements AsyncBackend. Idempotent: GetSandbox short-circuits before CreateSandbox.
-func (b *agentHarnessOpenShellBackend) EnsureAgentHarness(ctx context.Context, ah *v1alpha2.AgentHarness) (sandboxbackend.EnsureResult, error) {
-	if ah == nil {
-		return sandboxbackend.EnsureResult{}, fmt.Errorf("AgentHarness is required")
-	}
-
-	ctx, cancel := b.CallCtx(ctx)
-	defer cancel()
-	ctx = withAuth(ctx, b.cfg.Token)
-
+// findExistingSandbox is the GetSandbox/NotFound idempotency probe shared across backends. It
+// returns (res, true, nil) when the gateway already has a sandbox under this AgentHarness's
+// deterministic name, (zero, false, nil) when the sandbox does not yet exist, or
+// (zero, false, err) on any other gateway error. Callers must already have applied CallCtx and
+// withAuth to ctx.
+func (b *agentHarnessOpenShellBackend) findExistingSandbox(ctx context.Context, ah *v1alpha2.AgentHarness) (sandboxbackend.EnsureResult, bool, error) {
 	osCli := b.openShell()
 	if osCli == nil {
-		return sandboxbackend.EnsureResult{}, fmt.Errorf("openshell: OpenShell client is required (use Dial or non-nil OpenShellClients.OpenShell)")
+		return sandboxbackend.EnsureResult{}, false, fmt.Errorf("openshell: OpenShell client is required (use Dial or non-nil OpenShellClients.OpenShell)")
 	}
-
 	name := agentHarnessGatewayName(ah)
-
 	getResp, err := osCli.GetSandbox(ctx, &openshellv1.GetSandboxRequest{Name: name})
 	if err == nil && getResp != nil && getResp.GetSandbox() != nil {
 		handleID := sandboxBackendHandleID(getResp.GetSandbox())
 		return sandboxbackend.EnsureResult{
 			Handle:   sandboxbackend.Handle{ID: handleID},
 			Endpoint: endpointFor(b.cfg.GatewayURL, handleID),
-		}, nil
+		}, true, nil
 	}
 	if err != nil && status.Code(err) != codes.NotFound {
-		return sandboxbackend.EnsureResult{}, fmt.Errorf("openshell GetSandbox %s: %w", name, err)
+		return sandboxbackend.EnsureResult{}, false, fmt.Errorf("openshell GetSandbox %s: %w", name, err)
 	}
+	return sandboxbackend.EnsureResult{}, false, nil
+}
 
-	if b.preCreateSandbox != nil {
-		if err := b.preCreateSandbox(ctx, ah, b.kubeClient, b.clients); err != nil {
-			return sandboxbackend.EnsureResult{}, err
-		}
-	}
-
+// createSandbox builds the gateway create request via b.buildCreate and submits it. Callers must
+// already have applied CallCtx and withAuth to ctx and confirmed via findExistingSandbox that no
+// sandbox exists yet.
+func (b *agentHarnessOpenShellBackend) createSandbox(ctx context.Context, ah *v1alpha2.AgentHarness) (sandboxbackend.EnsureResult, error) {
 	req, unsupported := b.buildCreate(ah)
 	return b.CreateAgentHarnessSandbox(ctx, ah, req, unsupported)
 }
@@ -114,24 +95,4 @@ func (b *agentHarnessOpenShellBackend) GetStatus(ctx context.Context, h sandboxb
 // DeleteAgentHarness implements AsyncBackend.
 func (b *agentHarnessOpenShellBackend) DeleteAgentHarness(ctx context.Context, h sandboxbackend.Handle) error {
 	return b.DeleteAgentHarnessSandbox(ctx, h)
-}
-
-// OpenShellBackend implements AsyncBackend for spec.backend=openshell (generic
-// OpenShell sandbox provisioning without OpenClaw/Nemo bootstrap behavior).
-type OpenShellBackend struct {
-	*agentHarnessOpenShellBackend
-}
-
-var _ sandboxbackend.AsyncBackend = (*OpenShellBackend)(nil)
-
-// NewOpenShellBackend returns a backend for AgentHarness.spec.backend=openshell.
-func NewOpenShellBackend(kubeClient client.Client, clients *OpenShellClients, cfg Config, recorder record.EventRecorder) *OpenShellBackend {
-	return &OpenShellBackend{
-		agentHarnessOpenShellBackend: newAgentHarnessOpenShellBackend(
-			kubeClient, clients, cfg, recorder,
-			v1alpha2.AgentHarnessBackendOpenshell,
-			buildAgentHarnessOpenshellCreateRequest,
-			nil,
-		),
-	}
 }
